Chal #13 Wallet Miner: Complete
Call GnosisSafe's executeTransaction() with a EIP-712 signature.

Note: The GnosisSafe version (1.3.0) included in dependencies is not
compliant. The one we work with (0x34CfAC646f301356fAa8B21e94227e3583Fe3F5F)
does not use chain id (EIP-155). This has resulted in problems when we used the
utilities to sign the tx.